fix(avalonia): address #775 review — ShowError on failure, split blank/invalid-pointer guards (#773)
Copilot review on PR #775:
- View ExpandList_Click: use ShowError (not ShowInfo) on failure, matching
  the sibling NewAlloc_Click path so expansion failures aren't easy to miss.
- VM ExpandExitList: split the IsBlank guard ("use New") from the
  CurrentExitPointAddr==NOT_FOUND guard ("invalid/unset pointer; cannot
  expand") — NOT_FOUND with IsBlank=false means a garbage slot pointer where
  NewAlloc is hidden, so the "use New" message was misleading there.
- Test: cast the terminator-write index to int explicitly (matches the
  rowBase cast above); the uint+int expression already compiled (uint+int
  promotes to long; CI was green) but the cast is clearer and consistent.

All 28 MapExitPoint Avalonia tests pass.

+++ FEBuilderGBA.Avalonia.Tests/GapSweep/MapExitPointExpandTests.cs
@@ -83,7 +83,7 @@ public class MapExitPointExpandTests
             bytes[rowBase + 2] = 0x03;
             bytes[rowBase + 3] = 0x00;
         }
-        bytes[PerMapBase + rows * 4] = 0xFF; // terminator
+        bytes[(int)(PerMapBase + rows * 4)] = 0xFF; // terminator (explicit cast, matches rowBase above)
 
         // Explicit 0xFF free-space region so FindFreeSpace lands here.
         for (uint i = 0; i < 0x40000u; i++)
